// src/packages/ResultCode.d.ts
export declare enum ResultCode {
    OK = 0,
    CONNECTION_ERROR = 1,
    CONFIGURATION_ERROR = 2,
    FAIL = 3,
    INCORRECT_LOGIN_OR_PASSWORD = 4012,
    LESS_THAN_8 = 4013,
    NOT_HAS_LOWER_LETTERS = 4014,
    NOT_HAS_UPPER_LETTERS = 4015,
    NOT_HAS_DIGIT = 4016,
    NOT_HAS_SPECIAL_SIGNS = 4017,
    HAS_WHITE_SPACES = 4018,
    REGISTER_EXISTING_LOGIN = 4019,
    INCORRECT_TOKEN = 5001,
    EXPIRED_TOKEN = 5002
}